合并文件到solve_network_myopic
