+++ config.yaml
@@ -20,7 +20,7 @@ tmax: '-12-31 23:00'
 freq: '24h'
 
 add_aluminum: True
-iterative_optimization: False
+iterative_optimization: True
 aluminum_commitment: False
 aluminum_max_iterations: 5
 aluminum_convergence_tolerance: 0.01
