+++ scripts/solve_network_myopic.py
@@ -6,15 +6,22 @@
 
 import logging
 import re
-
+import copy
 import numpy as np
 import pandas as pd
 import pypsa
 import xarray as xr
+import os
 from _helpers import (
     configure_logging,
     override_component_attrs,
 )
+
+# 允许传递给PyPSA optimize的参数
+ALLOWED_OPTIMIZE_KWARGS = [
+    "solver_name", "solver_options", "extra_functionality",
+    "track_iterations", "min_iterations", "max_iterations"
+]
 
 logger = logging.getLogger(__name__)
 pypsa.pf.logger.setLevel(logging.WARNING)
@@ -248,7 +255,7 @@ def add_retrofit_constraints(n):
                 rhs = max_cap - retrofit_cap
                 n.model.add_constraints(lhs == rhs, name=f"Generator-coal-retrofit-{year}-{bus}")
 
-def extra_functionality(n, snapshots):
+def extra_functionality(n, snapshots, fixed_aluminum_usage=None):
     """
     Collects supplementary constraints which will be passed to ``pypsa.linopf.network_lopf``.
     If you want to enforce additional custom constraints, this is a good location to add them.
@@ -261,7 +268,377 @@ def extra_functionality(n, snapshots):
     if snakemake.wildcards.planning_horizons != "2020":
         add_retrofit_constraints(n)
 
-def solve_network(n, config, solving, opts="", **kwargs):
+def solve_aluminum_optimization(n, config, solving, opts="", nodal_prices=None, **kwargs):
+    """
+    电解铝最优运行问题求解
+    基于节点电价，运行以满足铝需求为约束的电解铝最优运行问题
+    """
+    set_of_options = solving["solver"]["options"]
+    solver_options = solving["solver_options"][set_of_options] if set_of_options else {}
+    solver_name = solving["solver"]["name"]
+    
+    # 不复制网络，而是直接使用传入的网络
+    # 找到所有电解铝相关的组件
+    aluminum_buses = n.buses[n.buses.carrier == "aluminum"].index
+    aluminum_smelters = n.links[n.links.carrier == "aluminum smelter"].index
+    aluminum_stores = n.stores[n.stores.carrier == "aluminum storage"].index
+    aluminum_loads = n.loads[n.loads.bus.isin(aluminum_buses)].index
+    
+    # 重新设置电解铝冶炼设备的参数
+    for smelter in aluminum_smelters:
+        n.links.at[smelter, 'committable'] = config['aluminum_commitment']
+        n.links.at[smelter, 'p_min_pu'] = config['aluminum']['al_p_min_pu'] if config['aluminum_commitment'] else 0
+    
+    # 移除所有非电解铝相关的组件
+    for component_type in ["Generator", "StorageUnit", "Store", "Link", "Load"]:
+        if component_type == "Store":
+            # 保留电解铝存储
+            other_stores = n.stores[~n.stores.index.isin(aluminum_stores)].index
+            n.mremove(component_type, other_stores)
+        elif component_type == "Link":
+            # 保留电解铝冶炼设备
+            other_links = n.links[~n.links.index.isin(aluminum_smelters)].index
+            n.mremove(component_type, other_links)
+        elif component_type == "Load":
+            # 保留电解铝负荷
+            other_loads = n.loads[~n.loads.index.isin(aluminum_loads)].index
+            n.mremove(component_type, other_loads)
+        else:
+            # 移除所有其他组件
+            n.mremove(component_type, n.df(component_type).index)
+    
+    # 移除aluminum和AC以外的节点
+    non_aluminum_buses = n.buses[(n.buses.carrier != "aluminum") & (n.buses.carrier != "AC")].index
+    n.mremove("Bus", non_aluminum_buses)
+    
+    # 确保虚拟carrier存在
+    if "virtual" not in n.carriers.index:
+        n.add("Carrier", "virtual")
+    
+    # 添加虚拟发电机来提供电力（基于节点电价）
+    for bus in n.buses.index:
+        if bus.endswith(" aluminum"):
+            # 这是电解铝节点，不需要虚拟发电机
+            continue
+            
+        # 根据节点类型确定边际成本
+        if bus in nodal_prices.columns:
+            # 如果该节点有对应的边际电价，使用该节点的电价
+            marginal_cost = nodal_prices[bus]
+            logger.info(f"为节点 {bus} 使用其对应的边际电价")
+            
+        n.add("Generator",
+            f"virtual_gen_{bus}",
+            bus=bus,
+            carrier="virtual",
+            p_nom=1e6,  # 大容量
+            marginal_cost=marginal_cost)  # 使用节点边际电价作为边际成本
+            
+    # 打印所有bus:
+    logger.info(f"所有bus: {n.buses.index}")
+    
+    # 求解电解铝优化问题
+    try:
+        # 只保留PyPSA支持的参数
+        optimize_kwargs = {k: v for k, v in kwargs.items() if k in ALLOWED_OPTIMIZE_KWARGS}
+        status, condition = n.optimize(
+            solver_name=solver_name,
+            extra_functionality=extra_functionality,
+            **solver_options,
+            **optimize_kwargs,
+        )
+        
+        logger.info("电解铝优化问题求解成功")
+        # 提取电解铝用能模式
+        aluminum_usage = n.links_t.p0[aluminum_smelters].copy()
+        return aluminum_usage
+            
+    except Exception as e:
+        logger.error(f"电解铝优化问题求解出错: {e}")
+        return None
+
+def solve_network_iterative(n, config, solving, opts="", max_iterations=10, convergence_tolerance=0.01, **kwargs):
+    """
+    电解铝迭代优化算法
+    1. 使用连续化电解铝模型求解，得到节点电价和目标函数值
+    2. 检查收敛性：比较步骤一的目标函数值变化
+    3. 基于节点电价，运行电解铝最优运行问题，得到新的电解铝用能模式
+    4. 固定电解铝用能，求解剩下的优化问题
+    5. 重复1-4，直到收敛
+    
+    收敛条件：步骤一目标函数变化的相对值小于convergence_tolerance（默认1%）
+    """
+    import time
+    
+    set_of_options = solving["solver"]["options"]
+    solver_options = solving["solver_options"][set_of_options] if set_of_options else {}
+    solver_name = solving["solver"]["name"]
+    cf_solving = solving["options"]
+    track_iterations = cf_solving.get("track_iterations", False)
+    min_iterations = cf_solving.get("min_iterations", 4)
+    max_transmission_iterations = cf_solving.get("max_iterations", 6)
+
+    skip_iterations = cf_solving.get("skip_iterations", False)
+    if not n.lines.s_nom_extendable.any():
+        skip_iterations = True
+        logger.info("No expandable lines found. Skipping iterative solving.")
+    
+    # 检查是否启用电解铝
+    if not config.get("add_aluminum", False):
+        logger.info("电解铝功能未启用，使用标准求解方法")
+        return solve_network_standard(n, config, solving, opts, **kwargs)
+    
+    logger.info("开始电解铝迭代优化算法")
+    
+    # 记录总开始时间
+    total_start_time = time.time()
+    
+    # 初始化电解铝用能模式和目标函数值
+    aluminum_usage = None
+    previous_objective = None
+    iteration = 0
+    final_network = None
+    
+    # 记录每次迭代的时间
+    iteration_times = []
+    
+    # 保存原始网络文件路径和overrides，用于重新加载
+    original_network_path = None
+    original_overrides = None
+    if hasattr(n, '_network_path'):
+        original_network_path = n._network_path
+    if hasattr(n, '_overrides_path'):
+        original_overrides = override_component_attrs(n._overrides_path)
+    
+    # 获取单节点参数
+    using_single_node = kwargs.get("using_single_node", False)
+    single_node_province = kwargs.get("single_node_province", "Shandong")
+    
+    while iteration < max_iterations:
+        iteration += 1
+        iteration_start_time = time.time()  # 记录每次迭代开始时间
+        
+        logger.info(f"开始第 {iteration} 次迭代")
+        
+        # 步骤1: 使用连续化电解铝模型求解，得到节点电价
+        logger.info("步骤1: 使用连续化电解铝模型求解")
+        
+        # 重新加载网络，确保网络状态一致
+        if original_network_path:
+            if original_overrides:
+                n_current = pypsa.Network(original_network_path, override_component_attrs=original_overrides)
+            else:
+                n_current = pypsa.Network(original_network_path)
+            
+            # 设置网络文件路径，用于后续可能的重新加载
+            n_current._network_path = original_network_path
+            if original_overrides:
+                n_current._overrides_path = kwargs.get("overrides_path", "data/override_component_attrs")
+
+        
+        # 重新应用网络准备
+        n_current = prepare_network(
+            n_current,
+            kwargs.get("solve_opts", {}),
+            using_single_node=using_single_node,
+            single_node_province=single_node_province
+        )
+        
+        # 设置配置
+        n_current.config = config
+        n_current.opts = opts
+        
+        # 重新创建电解铝冶炼设备，确保在步骤1中禁用电解铝启停约束
+        aluminum_smelters = n_current.links[n_current.links.carrier == "aluminum smelter"].index
+        
+        # 保存原始参数
+        smelter_params = {}
+        for smelter in aluminum_smelters:
+            smelter_params[smelter] = {
+                'bus0': n_current.links.at[smelter, 'bus0'],
+                'bus1': n_current.links.at[smelter, 'bus1'],
+                'carrier': n_current.links.at[smelter, 'carrier'],
+                'p_nom': n_current.links.at[smelter, 'p_nom'],
+                'p_nom_extendable': n_current.links.at[smelter, 'p_nom_extendable'],
+                'efficiency': n_current.links.at[smelter, 'efficiency'],
+                'start_up_cost': n_current.links.at[smelter, 'start_up_cost'],
+                'committable': n_current.links.at[smelter, 'committable'],
+                'p_min_pu': n_current.links.at[smelter, 'p_min_pu']
+            }
+        
+        # 移除现有的smelter
+        n_current.mremove("Link", aluminum_smelters)
+        
+        # 重新添加smelter，在步骤1中禁用电解铝启停约束
+        for smelter_name, params in smelter_params.items():
+            n_current.add("Link",
+                smelter_name,
+                bus0=params['bus0'],
+                bus1=params['bus1'],
+                carrier=params['carrier'],
+                p_nom=params['p_nom'],
+                p_nom_extendable=params['p_nom_extendable'],
+                efficiency=params['efficiency'],
+                start_up_cost=params['start_up_cost'],
+                committable=False,  # 在步骤1中禁用电解铝启停约束
+                p_min_pu=0  # 在步骤1中设置最小出力为0
+            )
+        
+        # 临时修改配置，禁用电解铝启停约束
+        original_commitment = config.get("aluminum_commitment", False)
+        config["aluminum_commitment"] = False
+        
+        # 如果有固定的电解铝用能，需要添加约束
+        if aluminum_usage is not None:
+            logger.info("根据电解铝用能模式设置动态约束")
+            # 根据电解铝用能模式动态设置约束
+            for smelter in aluminum_usage.columns:
+                if smelter in n_current.links.index:
+                    aluminum_power = aluminum_usage[smelter].values
+                                     
+                    # 根据用能模式设置约束
+                    for i, power in enumerate(aluminum_power):
+                        if power == 0:
+                            # 当用能为0时，固定p_set为0，p_min_pu为0
+                            n_current.links_t.p_set.at[n_current.snapshots[i], smelter] = 0
+                            n_current.links_t.p_min_pu.at[n_current.snapshots[i], smelter] = 0
+                        else:
+                            # 当用能不为0时，不固定p_set，但设置p_min_pu为最小出力比例
+                            # 清除p_set约束（设为NaN表示不约束）
+                            n_current.links_t.p_set.at[n_current.snapshots[i], smelter] = np.nan
+                            # 设置最小出力比例，可以根据实际需求调整
+                            min_pu = config['aluminum'].get('al_p_min_pu', 0.9)  # 默认30%最小出力
+                            n_current.links_t.p_min_pu.at[n_current.snapshots[i], smelter] = min_pu
+                    
+                    logger.info(f"为电解铝冶炼设备 {smelter} 设置动态约束")
+                    logger.info(f"  零出力时段数: {np.sum(aluminum_power == 0)}")
+                    logger.info(f"  非零出力时段数: {np.sum(aluminum_power > 0)}")
+        
+        # 求解网络
+        try:
+            # 只保留PyPSA支持的参数
+            optimize_kwargs = {k: v for k, v in kwargs.items() if k in ALLOWED_OPTIMIZE_KWARGS}
+            if skip_iterations:
+                status, condition = n_current.optimize(
+                    solver_name=solver_name,
+                    extra_functionality=extra_functionality,
+                    **solver_options,
+                    **optimize_kwargs,
+                )
+            else:
+                status, condition = n_current.optimize.optimize_transmission_expansion_iteratively(
+                    solver_name=solver_name,
+                    track_iterations=track_iterations,
+                    min_iterations=min_iterations,
+                    max_iterations=max_transmission_iterations,
+                    extra_functionality=extra_functionality,
+                    **solver_options,
+                    **optimize_kwargs,
+                )
+            
+            logger.info("网络求解成功，获得节点电价")
+            
+            # 获取当前目标函数值
+            current_objective = None
+            if hasattr(n_current, 'objective'):
+                current_objective = n_current.objective
+            elif hasattr(n_current, 'model') and hasattr(n_current.model, 'objective_value'):
+                current_objective = n_current.model.objective_value
+            
+            # 提取当前迭代的节点电价（用于电解铝优化）
+            current_nodal_prices = None
+            if hasattr(n_current, 'buses_t') and hasattr(n_current.buses_t, 'marginal_price'):
+                # 获取所有carrier为"AC"的电力节点的边际电价
+                electricity_buses = n_current.buses[n_current.buses.carrier == "AC"].index
+                if len(electricity_buses) > 0:
+                    # 使用所有AC节点的边际电价
+                    current_nodal_prices = n_current.buses_t.marginal_price[electricity_buses]
+                    logger.info(f"提取节点电价: 共 {len(electricity_buses)} 个AC节点")
+                    logger.info(f"节点列表: {list(electricity_buses)}")
+                else:
+                    logger.warning("未找到carrier为AC的节点，无法提取节点电价")
+            
+        except Exception as e:
+            logger.error(f"网络求解出错: {e}")
+            break
+        
+        # 步骤2: 检查收敛性 - 基于步骤一的目标函数变化的相对值
+        if previous_objective is not None and current_objective is not None:
+            # 计算目标函数变化的相对值
+            objective_change = abs(current_objective - previous_objective)
+            relative_change = objective_change / abs(previous_objective) if abs(previous_objective) > 1e-10 else float('inf')
+            
+            logger.info(f"步骤一目标函数变化统计:")
+            logger.info(f"  当前目标函数值: {current_objective:.6e}")
+            logger.info(f"  上次目标函数值: {previous_objective:.6e}")
+            logger.info(f"  绝对变化: {objective_change:.6e}")
+            logger.info(f"  相对变化: {relative_change:.6f} ({relative_change*100:.2f}%)")
+            logger.info(f"  收敛阈值: {convergence_tolerance:.6f} ({convergence_tolerance*100:.2f}%)")
+            
+            if relative_change < convergence_tolerance:
+                logger.info(f"算法收敛，在第 {iteration} 次迭代后停止")
+                logger.info(f"目标函数相对变化 {relative_change:.6f} ({relative_change*100:.2f}%) < 收敛阈值 {convergence_tolerance:.6f} ({convergence_tolerance*100:.2f}%)")
+                final_network = n_current
+                break
+        else:
+            # 第一次迭代，保存目标函数值用于下次比较
+            if current_objective is not None:
+                previous_objective = current_objective
+        
+        # 步骤3: 基于节点电价，运行电解铝最优运行问题
+        logger.info("步骤3: 运行电解铝最优运行问题")
+        
+        # 恢复电解铝启停约束
+        config["aluminum_commitment"] = True
+        
+        # 求解电解铝优化问题，传递节点电价
+        new_aluminum_usage = solve_aluminum_optimization(n_current, config, solving, opts, nodal_prices=current_nodal_prices, **kwargs)
+        
+        if new_aluminum_usage is None:
+            logger.error("电解铝优化问题求解失败")
+            break
+        
+        # 更新电解铝用能和目标函数值
+        aluminum_usage = new_aluminum_usage
+        previous_objective = current_objective
+        final_network = n_current
+        
+        # 记录本次迭代时间
+        iteration_time = time.time() - iteration_start_time
+        iteration_times.append(iteration_time)
+        
+        logger.info(f"第 {iteration} 次迭代完成，耗时: {iteration_time:.2f} 秒")
+    
+    # 恢复原始配置
+    config["aluminum_commitment"] = original_commitment
+    
+    # 计算总时间
+    total_time = time.time() - total_start_time
+    
+    if iteration >= max_iterations:
+        logger.warning(f"达到最大迭代次数 {max_iterations}，算法未完全收敛")
+    
+    # 输出时间统计
+    logger.info(f"\n=== 迭代时间统计 ===")
+    logger.info(f"总迭代次数: {iteration}")
+    logger.info(f"总耗时: {total_time:.2f} 秒 ({total_time/60:.2f} 分钟)")
+    if iteration_times:
+        logger.info(f"平均每次迭代耗时: {np.mean(iteration_times):.2f} 秒")
+        logger.info(f"最快迭代耗时: {np.min(iteration_times):.2f} 秒")
+        logger.info(f"最慢迭代耗时: {np.max(iteration_times):.2f} 秒")
+        logger.info(f"迭代时间详情:")
+        for i, t in enumerate(iteration_times, 1):
+            logger.info(f"  第{i}次迭代: {t:.2f} 秒")
+    
+    logger.info(f"电解铝迭代优化算法完成，共进行 {iteration} 次迭代")
+    
+    # 返回最终的网络结果
+    return final_network
+
+def solve_network_standard(n, config, solving, opts="", **kwargs):
+    """
+    标准网络求解方法（非迭代）
+    """
     set_of_options = solving["solver"]["options"]
     solver_options = solving["solver_options"][set_of_options] if set_of_options else {}
     solver_name = solving["solver"]["name"]
@@ -279,14 +656,16 @@ def solve_network(n, config, solving, opts="", **kwargs):
         skip_iterations = True
         logger.info("No expandable lines found. Skipping iterative solving.")
     
-    # 尝试使用默认参数求解
+    # 使用标准参数求解
     try:
+        # 只保留PyPSA支持的参数
+        optimize_kwargs = {k: v for k, v in kwargs.items() if k in ALLOWED_OPTIMIZE_KWARGS}
         if skip_iterations:
             status, condition = n.optimize(
                 solver_name=solver_name,
                 extra_functionality=extra_functionality,
                 **solver_options,
-                **kwargs,
+                **optimize_kwargs,
             )
         else:
             status, condition = n.optimize.optimize_transmission_expansion_iteratively(
@@ -296,133 +675,14 @@ def solve_network(n, config, solving, opts="", **kwargs):
                 max_iterations=max_iterations,
                 extra_functionality=extra_functionality,
                 **solver_options,
-                **kwargs,
+                **optimize_kwargs,
             )
 
-        # 处理不同的求解状态
-        if status == "ok":
-            if "optimal" in condition or "time_limit" in condition:
-                logger.info(f"求解成功，状态: {status}, 条件: {condition}")
-                if "time_limit" in condition:
-                    logger.warning("求解达到时间限制，返回次优解")
-            else:
-                logger.warning(f"求解状态 '{status}' with termination condition '{condition}'")
-        elif status == "warning":
-            if "time_limit" in condition:
-                logger.warning("求解超时，但获得了次优解")
-                # 次优解仍然可以使用
-            else:
-                logger.warning(f"求解警告，状态: {status}, 条件: {condition}")
-        else:
-            logger.warning(f"求解状态 '{status}' with termination condition '{condition}'")
-            
-        if "infeasible" in condition:
-            raise RuntimeError("Solving status 'infeasible'")
+        logger.info("标准求解成功")
 
     except Exception as e:
-        error_msg = str(e).lower()
-        # 检查是否是数值问题
-        if any(keyword in error_msg for keyword in ['numerical', 'infeasible', 'unbounded', 'barhomogeneous']):
-            logger.warning(f"遇到数值问题: {e}")
-            logger.info("尝试使用保守的求解器参数...")
-            
-            # 尝试使用保守参数
-            try:
-                conservative_options = solving["solver_options"].get("conservative", {})
-                if conservative_options:
-                    logger.info("使用保守求解器参数重新求解...")
-                    if skip_iterations:
-                        status, condition = n.optimize(
-                            solver_name=solver_name,
-                            solver_options=conservative_options,
-                            extra_functionality=extra_functionality,
-                            **kwargs,
-                        )
-                    else:
-                        status, condition = n.optimize.optimize_transmission_expansion_iteratively(
-                            solver_name=solver_name,
-                            track_iterations=track_iterations,
-                            min_iterations=min_iterations,
-                            max_iterations=max_iterations,
-                            extra_functionality=extra_functionality,
-                            **conservative_options,
-                            **kwargs,
-                        )
-                    
-                    # 处理保守参数求解的结果
-                    if status == "ok":
-                        if "optimal" in condition or "time_limit" in condition:
-                            logger.info(f"保守参数求解成功，状态: {status}, 条件: {condition}")
-                            if "time_limit" in condition:
-                                logger.warning("保守参数求解也达到时间限制，返回次优解")
-                        else:
-                            logger.warning(f"保守参数求解状态 '{status}' with termination condition '{condition}'")
-                    elif status == "warning":
-                        if "time_limit" in condition:
-                            logger.warning("保守参数求解超时，但获得了次优解")
-                        else:
-                            logger.warning(f"保守参数求解警告，状态: {status}, 条件: {condition}")
-                    else:
-                        logger.warning(f"保守参数求解状态 '{status}' with termination condition '{condition}'")
-                        
-                    if "infeasible" in condition:
-                        raise RuntimeError("保守参数求解状态 'infeasible'")
-                else:
-                    logger.error("未找到保守求解器参数配置")
-                    raise e
-            except Exception as e2:
-                logger.error(f"保守参数求解也失败: {e2}")
-                
-                # 尝试第三级超时回退选项
-                try:
-                    timeout_fallback_options = solving["solver_options"].get("timeout_fallback", {})
-                    if timeout_fallback_options:
-                        logger.info("使用超时回退参数重新求解...")
-                        if skip_iterations:
-                            status, condition = n.optimize(
-                                solver_name=solver_name,
-                                solver_options=timeout_fallback_options,
-                                extra_functionality=extra_functionality,
-                                **kwargs,
-                            )
-                        else:
-                            status, condition = n.optimize.optimize_transmission_expansion_iteratively(
-                                solver_name=solver_name,
-                                track_iterations=track_iterations,
-                                min_iterations=min_iterations,
-                                max_iterations=max_iterations,
-                                extra_functionality=extra_functionality,
-                                **timeout_fallback_options,
-                                **kwargs,
-                            )
-                        
-                        # 处理超时回退求解的结果
-                        if status == "ok":
-                            if "optimal" in condition or "time_limit" in condition:
-                                logger.info(f"超时回退求解成功，状态: {status}, 条件: {condition}")
-                                if "time_limit" in condition:
-                                    logger.warning("超时回退求解也达到时间限制，返回次优解")
-                            else:
-                                logger.warning(f"超时回退求解状态 '{status}' with termination condition '{condition}'")
-                        elif status == "warning":
-                            if "time_limit" in condition:
-                                logger.warning("超时回退求解超时，但获得了次优解")
-                            else:
-                                logger.warning(f"超时回退求解警告，状态: {status}, 条件: {condition}")
-                        else:
-                            logger.warning(f"超时回退求解状态 '{status}' with termination condition '{condition}'")
-                            
-                        if "infeasible" in condition:
-                            raise RuntimeError("超时回退求解状态 'infeasible'")
-                    else:
-                        logger.error("未找到超时回退求解器参数配置")
-                        raise e2
-                except Exception as e3:
-                    logger.error(f"超时回退求解也失败: {e3}")
-                    raise e3
-        else:
-            # 如果不是数值问题，直接抛出原始异常
-            raise e
+        logger.error(f"标准求解失败: {e}")
+        raise e
 
     # Store the objective value from the model (兼容性处理)
     try:
@@ -440,12 +700,12 @@ def solve_network(n, config, solving, opts="", **kwargs):
 if __name__ == '__main__':
     if 'snakemake' not in globals():
         from _helpers import mock_snakemake
-        snakemake = mock_snakemake('solve_network_myopic',
+        snakemake = mock_snakemake('solve_network_myopic-1',
                                    opts='ll',
                                    topology='current+Neighbor',
-                                   pathway='exponential175',
+                                   pathway='linear2050',
                                    co2_reduction='0.0',
-                                   planning_horizons="2025")
+                                   planning_horizons="2050")
 
     configure_logging(snakemake)
 
@@ -462,6 +722,11 @@ if __name__ == '__main__':
         n = pypsa.Network(snakemake.input.network, override_component_attrs=overrides)
     else:
         n = pypsa.Network(snakemake.input.network)
+    
+    # 设置网络文件路径，用于迭代优化中的网络重新加载
+    n._network_path = snakemake.input.network
+    if "overrides" in snakemake.input.keys():
+        n._overrides_path = snakemake.input.overrides
 
     n = prepare_network(
         n,
@@ -470,13 +735,52 @@ if __name__ == '__main__':
         single_node_province=snakemake.params.single_node_province
     )
 
-    n = solve_network(
-        n,
-        config=snakemake.config,
-        solving=snakemake.params.solving,
-        opts=opts,
-        log_fn=snakemake.log.solver,
-    )
+    # 检查是否启用电解铝迭代优化
+    if snakemake.params.iterative_optimization:
+        # 获取迭代优化参数
+        max_iterations = snakemake.config.get("aluminum_max_iterations", 10)
+        convergence_tolerance = snakemake.config.get("aluminum_convergence_tolerance", 0.01)
+        
+        logger.info(f"启用电解铝迭代优化算法")
+        logger.info(f"最大迭代次数: {max_iterations}")
+        logger.info(f"收敛阈值: {convergence_tolerance}")
+        
+        # 准备传递给迭代函数的参数
+        iteration_kwargs = {
+            "log_fn": snakemake.log.solver,
+            "solve_opts": solve_opts,
+            "using_single_node": snakemake.params.using_single_node,
+            "single_node_province": snakemake.params.single_node_province
+        }
+        
+        # 如果有overrides，也传递
+        if "overrides" in snakemake.input.keys():
+            iteration_kwargs["overrides"] = overrides
+        
+        # 使用迭代优化算法
+        n = solve_network_iterative(
+            n,
+            config=snakemake.config,
+            solving=snakemake.params.solving,
+            opts=opts,
+            max_iterations=max_iterations,
+            convergence_tolerance=convergence_tolerance,
+            **iteration_kwargs,
+        )
+    else:
+        logger.info("使用标准求解方法")
+        # 使用标准求解方法
+        n = solve_network_standard(
+            n,
+            config=snakemake.config,
+            solving=snakemake.params.solving,
+            opts=opts,
+            log_fn=snakemake.log.solver,
+        )
+
+    # 确保输出目录存在
+    output_dir = os.path.dirname(snakemake.output.network_name)
+    os.makedirs(output_dir, exist_ok=True)
 
     #n.meta = dict(snakemake.config, **dict(wildcards=dict(snakemake.wildcards)))
     n.links_t.p2 = n.links_t.p2.astype(float)
@@ -484,4 +788,7 @@ if __name__ == '__main__':
     if hasattr(n, 'links_t') and hasattr(n.links_t, 'p3'):
         # Convert DataFrame to numeric, handling any non-numeric values
         n.links_t.p3 = n.links_t.p3.apply(pd.to_numeric, errors='coerce').fillna(0.0)
+    
+    # 导出结果
     n.export_to_netcdf(snakemake.output.network_name)
+    logger.info(f"结果已保存到: {snakemake.output.network_name}")
